Provide image loading state in MetaImage via a publisher (#35)
Description
We expose a publisher with image loading state. Initial state is none and then it is updated to success for local image or success/error with remote images.

Motivation and Context
Using the onSuccess / onError publishers in the image flow, it is possible to provide alternate images to load. There was no way, however, to know wether the global image was loaded or not.

Using that technique, we're now able to take action on images errors by showing an alternate view for instance.