This fixes an issue where initial value of user input would immediately be replaced by empty string. (#38)
Description
This fixes an issue where initial value of user input would immediately be replaced by empty string.

What happened is that the text watcher would be triggered with an empty string by input type being set. We were then losing initial value.

This kind of rebound also happened with userInput observe but this time the the identical value which was less important.
So we now unset the watcher before changing the value, and then set it back again.

Motivation and Context
We couldn't provide an initial value for our input fields.