ListViewModel (#52)
Description
This adds a new component to the library: ListViewModel
The list view model comes with ListItemViewModel which is an item that can be included in a list. You can subclass ListItemViewModel and implement comparableId and optionnally isTheSame and haveTheSameContent which will be used on the different platforms for diff calculations.

On the android side, the LifecycleAdapter and ViewModelAdapter can be subclassed and support single item type as well as multiple item types.
LifecycleAdapter provides a way to bind lifecycle aware components. The item lifecycle is bound to an recyclerview item lifecycle such as when the view is detached or recycled from the recycling view it will trigger ON_STOP and ON_DESTROY events which is used to stop listening to LiveData objects.

On the iOS side, there might be some additional work to be done for the binding, I'm not really sure for this some someone could contribute with a convenience binding.

Motivation and Context
This is a component that was needed in our application and we found out that it helped reduce code boilerplate quite a lot in most simple cases and was customizable in complex ones.