Update to ktor 1.5.3 (#47)
Description
Since updating our project to ktor 1.5.X received empty http error message body.
The issue was introduced in ktor 1.5.0 it seems with this commit: ktorio/ktor@4798b0f

What happens is that the body is read to set the exception message without "saving" it to be re-used later. So when we were reading it in trikot.http, it was empty as it was already consumed. The proper fix (and I think I will submit a PR to ktor) would be to save the body so it can be read again after that.

See:
https://github.com/ktorio/ktor/blob/e425a2885f8e4ff0158c38c6595f2f2169fa3c9a/ktor-client/ktor-client-core/common/src/io/ktor/client/features/DefaultResponseValidation.kt#L44

The workaround in our case is to disable that feature and since we do not require the response validation, it makes no change for us (http error status codes are handled in trikot.http by another mecanism).