Android reflection based view model provider factory (#98)
Description
Add view model provider factory which uses reflection to load view model (controllers) and allows to pass string parameters as argument to fragment or activity.

We've chosen JSON to pass arguments to view models cross platform hence the choise of string parameter here. We could add other type of cross platform serialization format if need be.

Motivation and Context
This allows creating view model on android with less boilerplate than having to instanciate them manually by providing a custom ViewModelFactory for each different type. It comes with a slight performance overhead.

How Has This Been Tested?
Unit tested with roboelectric and battle tested in an app.

Note
This is a more generic version of what we used in another project (cudos to @mathieularue for the original idea and implementation).