Exponential backoff policy (#104)
📖 Description
I also added an extension to retry with backOff, it basically combines retryWhen, switchMap and timer. I was hesistant to add it here but it should be a good starting point.
For more advanced usage (retrying on certain errors, or only when network is available) the BackoffStrategy will be available standalone (it is actually my use case).

💭 Motivation and Context
This should be useful with http requests for instance to retry automatically and provide custom backoff strategy to reduce server load.