Gradle.kts migration and dependencies refresh (#28)
📖 Description
gradle.kts is becoming the standard and as it does, more pluggins are exposing nice DSLs. For this reason I think it's a good time to migrate the boilerplate to this.

👷 Work done
Migrate patron to gradle.kts
Extract libs and versions to buildSrc
Keep using gradle properties for trikot dependencies, as they still have to be synced with cocoapod spec.
Update to kotlin 1.5 gradle 7 and latest trikot.