Fix double binding with ViewModelAdapter (#103)
Description
Fix issue in ViewModelAdapter where binding.invalidateAll() would cause binding to be executed twice for each list item. Not sure why we added that in the first place (must have been for a good reason), but with the latest version of androidx.lifecycle it does not seem to be required.

Fix issue in sample project, we were leaking fragments and activities due to navigation delegate being a strong reference and never being cleared. Weak atomic reference is the way to go for navigation delegates on all platforms.

Bit of refactor to use observe extension directly.

Motivation and Context
We were investigating an issue where our binding would be executed twice in a recycler view using the ViewModelAdapter. Found out the culprit was a invalidate call on the binding. While testing I found out memory leaks in sample project.