Once a view holder is detached and it's lifecycle is destroyed. We need to invalidate its bindings so that it re-executes them and uses the new lifecycle owner. (#105)
Description
Once a view holder is detached and its lifecycle is destroyed. We need to invalidate its bindings so that it re-executes them and uses the new lifecycle owner.

Motivation and Context
#103 was introduced to fix a double binding issue when the view was never bound before. It broke the behavior when a view is detached and re-attached, the binding wasn't executed, leading to updates to the viewmodel not updating the view.

How Has This Been Tested?
The binding is done only once on the first bind, and properly called after a detach / attach.