Bugfix - SwitchMap inner publisher cancellation order (#111)
📖 Description
This fixes an issue in SwitchMap. The cancel on the previous publisher was done AFTER subscribing to the new publisher.

According to the RX Spec, cancellation on previous publisher must take place before the new publisher subscription.
See https://github.com/ReactiveX/RxJava/blob/3.x/src/main/java/io/reactivex/rxjava3/internal/operators/observable/ObservableSwitchMap.java for details about implementation.

With a combinaison of RefreshablePublisher that our refresh to be triggered twice in a row (once in refresh and once in onFirstSubscription). Since we now cancel before refresh is called, only the one in onFirstSubscription is left.

💭 Motivation and Context
It caused issues combined to other types of Publishers.

🧪 How Has This Been Tested?
Unit tested and it fixes our issue where we had double execution of our refresh block.