[Kword] Dynamic I18N and support multiple json paths (#48)
Description
Introduce a new DynamicI18N concept which expose strings via Publishers. It is implemented via a MultiLanguageI18N to react on language changes. When the language switches all publishers will emit the localized string in the new language.

Allow loading json strings from multiple sources (paths) via the basePaths parameter.

Remove code duplication between platform string loaders by moving part of the logic to KwordLoader in the common code.

Test utils from Trikot Streams project was extracted so they can be used from outside of it.

Motivation and Context
Dynamic i18N was required as we wanted to switch strings without reloading the view.
basePaths was required as we needed to have multiple translations files to be merged into a single i18N and allow overrides.
How Has This Been Tested?
This has been living in part as an separate library for quite a while for now. The refactor was also tested in externally.