[VMD-Flow] Backport and fix some functionalities (#97)
Description
Re-activate sample projects in config, previously it was de-activated due a build issue on jenkins, I want to see how it does on github actions.
Add interactionSource and indication params to VMDButton, this was added in VMD but not VMD-Flow
Add support for transformText and unformatText to VMDTextField, this was added in VMD but not VMD-Flow
Fix issues with onReturnKeyTap in VMDTextField. It should not have been a StateFlow but a SharedFlow and it was still collecting flow after action was triggered.