[DataSourceFlow] Change delete/clear cache cancellation behavior (#114)
Description
Previously we were cancelling the cached data flow job when invalidating the cache via delete or clear.

That had the side effect to cancel all pending reads as well if both were launched at the same time (since the Cancellation did bubble up to the parent scope).

We found that this behavior was not really what we wanted to achieve. Now we no longer cancel the child job so that concurrent reads can continue. We also now start by clearing the cache before clearing the upstream so that local cache is cleared faster.

Motivation and Context
In some cases we did call clear and followed quickly by read. The clear did take some time to complete (due to IO with a file datasource) and read started while clear was running. Once clear completed, it cancelled the job, and the read was cancelled because of that.