Update to kotlin 1.8.10 and remove support for old memory model (#147)
Description
This updates to kotlin 1.8.10. In that version, calling freeze now cause compilation errors.
We are now dropping support for old MM

We are also updating various libraries.

Ktlint was updated so this is why there is a lot of code style changes.

Motivation and Context
Old MM has become obsolete in favor of the new MM.

How Has This Been Tested?
I tested it in another project to test ABI compatibility.

Types of changes
 Bug fix (non-breaking change which fixes an issue)
 New feature (non-breaking change which adds functionality)
 Breaking change (fix or feature that would cause existing functionality to change)
Changes are ABI compatible. Old freezing utils were deprecated and will just do nothing.
This is however breaking in the sense that we drop support for old memory model.

New android lifecycle library also as a breaking change but that's not on us.