[VMD] Fix image sampling on android (#151)
Description
Since we are using coil rememberAsyncImagePainter instead of AsyncImage, we are losing a few built-in features such as sampling in the image request decoding.

This passes size and scale parameters in the image request which will later be used to calculate the sample size. The size is a ConstraintsSizeResolver which will passed as a modifier for contraints to be resolved. This is a similar strategy employed in coil with AsyncImage.

NOTE: The Image is now nested within a Box and takes a block with a BoxScope as a receiver. If you are using that, you might want to switch your placeholder modifiers to match parent Box.

Motivation and Context
When large source images are provided, the image decoding could be slower and take more memory. This was due to the whole image being loaded even though the target size was be significantly smaller.

How Has This Been Tested?
Tested both in sample app and in another app.