When complex placeholder is not required, rely on AsyncImage. (#155)
Description
When complex placeholder is not required, rely on AsyncImage.

Using AsyncImage from coil is more efficient at measures and drawing. When there's no need to use a complex placeholder, we use this instead of having our own rendering with rememberAsyncImagePainter.

The placeholder is handled with a painter.

Motivation and Context
We're profiling and testing VMDImage with different Constraints. Turns out AsyncImage handles most of the constraints out of the box and is generally more efficient than we we currently have in VMD.

How Has This Been Tested?
Tested in sample app and in another app.

Changes should be ABI compatible and no changes should be needed to use the most efficient binding.