Update kotlin poet to latest (#175)
Description
Fix issue with kotlin poet recent versions which no longer accept override modifier on the constructor parameter, which is legit. We move it to the property which result in the same generated code.

Motivation and Context
Fixes #174

How Has This Been Tested?
Was able to reproduce locally with kotlin poet latest version, validated the fix too.