Assign Role.Button semantics by default for VMDButton and allow customization (#184)
Description
Assign Role.Button semantics by default for VMDButton and allow customization

Motivation and Context
On android, VMDButton implementation uses a Box and a clickable modifier. By default the role on the clickable modifier is null (no role). This will give Button role to all VMDButton by default and allow to override it if needed instead of the contrary (needing to set Button role each time).

How Has This Been Tested?
In sample