[VMD] Support edge to edge + switch to m3 in samples (#239)
Description
This adds support for edge to edge in VMD sample. Also switches a few material imports for m3 in those samples so default theme changed a bit too.

Motivation and Context
edge-to-edge was not properly supported yet in VMD sample and since now it's pretty much mandatory, it lead to content displaying behind status bar.